requirements.txt
PyYAML~=6.0.2
opencv-python~=4.10.0.84
numpy~=2.1.3
matplotlib~=3.10.0rc1
circle-fit~=0.2.1
scipy~=1.14.1
scikit-image~=0.24.0
tqdm~=4.67.1
networkx~=3.4.2
prettytable~=3.14.0
Cython~=3.0.12
setuptools~=75.8.0
pykdtree~=1.4.1